+++ src/platform/forms-system/src/js/patterns/minimal-header/minimalHeader.unit.spec.js
@@ -1,3 +1,4 @@
+import { expect } from 'chai';
 import sinon from 'sinon';
 import { afterEach } from 'mocha';
 import { isMinimalHeaderApp, isMinimalHeaderPath } from '.';
